Projeto inicia configuado fonts
